refactor: move files around to get a better view
