+++ src/Marvin.Cache.Headers/Domain/CacheLocation.cs
