+++ src/Marvin.Cache.Headers/Domain/ETag.cs
