+++ src/Marvin.Cache.Headers/Domain/ETagType.cs
