+++ src/Marvin.Cache.Headers/Domain/RequestKey.cs
@@ -1,4 +1,7 @@
-﻿namespace Marvin.Cache.Headers
+﻿// Any comments, input: @KevinDockx
+// Any issues, requests: https://github.com/KevinDockx/HttpCacheHeaders
+
+namespace Marvin.Cache.Headers
 {
     using System.Collections.Generic;
