+++ src/Marvin.Cache.Headers/Interfaces/IValidationValueStore.cs
@@ -7,9 +7,6 @@ namespace Marvin.Cache.Headers.Interfaces
 {
     public interface IValidationValueStore
     {
-        Task<ValidationValue> GetAsync(string key);
-        Task SetAsync(string key, ValidationValue eTag);
-
         Task<ValidationValue> GetAsync(RequestKey key);
         Task SetAsync(RequestKey key, ValidationValue eTag);
     }
